+++ intelmq/bots/collectors/mail/lib.py
@@ -19,7 +19,7 @@ class MailCollectorBot(CollectorBot):
             raise MissingDependencyError("imbox")
 
         if getattr(self.parameters, 'attach_unzip', None) and not self.extract_files:
-            self.parameters.extract_files = True
+            self.extract_files = True
             self.logger.warning("The parameter 'attach_unzip' is deprecated and will "
                                 "be removed in version 4.0. Use 'extract_files' instead.")